requirements.txt
Django~=1.11.0
djangorestframework
mysqlclient
Pillow